XIVY-2629 CacheAllRemoveUnusedAdvisor uses TTL for cache invalidation
To simplify and speed up the invalidation algorithm and also make it
easier to explain it to end users the following changes are made to the
cache invalidation of the CacheAllRemoveUnusedAdvisor:

The Cache advisor uses now a time to live (TTL) and time to idle (TTI)
parameter for cache invalidation. If an object was cached for more
seconds than configured in the timeToLive parameter it will be
invalidated. If an object was not used for more seconds than configured
in timeToIdle it will be invalidated. 0 means disabled.

In case too much objects are in the cache it uses now only the last
usage of the cached object to decide which objects are invalidated.

Remove getUsageCount from ICacheUsage and all its implementors. This
method is no longer used by CacheAllRemoveUnusedAdvisor.

Add migration notes for the changed configuration and MBean attribute

Synch ivy.cache.properties with the real caches properties

Do not cache all security member in memory to avoid out of memory