XIVY-6318 Upgrade Windows JRE from 11.0.9.1 to 11.0.13
Upgrade bundled JRE for Windows from AdoptOpenJdk 11.0.9.1 to
Adoptium/Temurin 11.0.13

OpenJDK 11.0.13 does not work with macOS because of JDK-8270216. This
bug causes UI deadlocks if you try to open a process in the Designer.
We therefore deliver Adoptium/Temurin 11.0.3 only on Windows.
For macOS we stick to AdoptOpenJdk 11.0.9.1 till 11.0.14 is released
which should fix the bug.