XIVY-12891 Configurator must only handle the channel specific settings
The configurator was introduced for the identity provider specific settings.
It was not used for the settings which are shared by all identity providers.
This behavior will be streamlined now for notificaiton channels.

If we want that the configuration also handles the shared settings, then
someone needs to implement that the same way also for the identity providers.

If we don't have the same approach for all features, we will add a lot of hacks
for example in engine cockpit to build dynamic UIs.