Streamline codebase by renaming transaction to tx
There was code which has used transaction and code
which has used tx as term for a database transaction.

With this change, the term has been streamlined to tx.
tx is a common abbreviation for transaction and is much
shorter. It's now much more pleasent to work in the database
layer and all layers which are heavily using our own
persistence framework.